# src/main/resources/v1/eric-oss-file-notification-enm-stub-openapi.yaml
#
# COPYRIGHT Ericsson 2021
#
#
#
# The copyright to the computer program(s) herein is the property of
#
# Ericsson Inc. The programs may be used and/or copied only with written
#
# permission from Ericsson Inc. or in accordance with the terms and
#
# conditions stipulated in the agreement/contract under which the
#
# program(s) have been supplied.
#

openapi: "3.0.2"
info:
  x-api-id: 3a3521d6-c070-4e75-9bc3-5ccb8e1d61c7
  x-audience: external-public
  description: "Micro Service Chassis"
  version: "0.0.1"
  title: "Micro Service Chassis"
  license:
    name: "COPYRIGHT Ericsson 2021"
    url: "https://www.ericsson.com/en/legal"
  termsOfService: "https://www.ericsson.com/en/legal"
  contact:
    email: "Team Quaranteam<PDLDUNQUAR@pdl.internal.ericsson.com>"
tags:
  - name: "sample"
    description: "Sample tag for demo purposes"
paths:
    /v1/sample:
      get:
        tags:
          - "sample"
        summary: "Returns a static response"
        operationId: "sample"
        responses:
          "200":
            description: "OK"
            content:
              application/json:
                schema:
                  type: "string"
          "404":
            description: "Not Found"
          default:
            description: "Internal Server Error"
    /login:
      post:
        tags:
          - "sample"
        summary: "Returns a static response"
        operationId: "login"
        responses:
          "200":
            description: "OK"
            content:
              application/json:
                schema:
                  type: "string"
          "404":
            description: "Not Found"
          default:
            description: "Internal Server Error"
    /file/v1/files:
      get:
        tags:
          - "sample"
        summary: "Returns a list file notification DTO-s"
        operationId: "getFiles"
        parameters:
          - in: query
            name: filter
            schema:
              type: "string"
            description: "Filter query"
          - in: query
            name: select
            schema:
              type: "string"
            description: "Select query"
          - in: query
            name: offset
            schema:
              type: "string"
            description: "Offset query"
          - in: query
            name: orderBy
            schema:
              type: "string"
            description: "Order by query"
        responses:
          "200":
            description: "OK"
            content:
              application/hal+json:
                schema:
                  type: "object"
          "404":
            description: "Not Found"
          default:
            description: "Internal Server Error"